[#170390872] XJ Program Editor (Stepmatic) Mk3 UI (#19)
  - **Globalheader** has logo and currently logged in user avatar
  - **Pattern** Edit, Create, Delete
  - **Program Memes** displayed and may be created/deleted
  - **Program** attributes edited in place
  - **Program** is saved. serializing program and posting changes to API
  - **ProgramNavigator** as header
  - **Sequence Binding Memes** displayed in sequence binding editor, can be added or removed from any sequence binding
  - **Sequence Bindings** displayed in horizontal grid grouped by offset, create new sequence binding in offset, remove any sequence binding
  - **Sequence Chord** show, create, edit, delete
  - **Sequence Pattern Event** show, create, edit, delete
  - **SequenceSelector** selects active sequence
  - **SequenceSelector/TimelineNavigator** bar is always fixed to top of view; stacked VoiceEditor area scrolls
  - **TimelineNavigator** displays timeline based on shared state of timeline size and divisions
  - **TimelineNavigator** selects divisions (1, 1/2, 1/4, 1/8, 1/16)
  - **Track** Edit, Create, Delete
  - **TrackEditor** shows grid based on timeline navigation divisions
  - **Voice** Edit, Create, Delete
  - **VoiceEditor** can rename the currently selected Pattern
  - **VoiceEditor** Row for each **Voice**
  - **Workspace Switcher** to switch between Sequence Bindings Editor and the Voice/Track/Pattern editor
  - [react-popup](https://react-popup.elazizi.com/) action menus to delete/clone/move, modal dialogs to handle warnings/confirms/inputs
  - Adds new **Sequence**
  - After deleting a Sequence remove it from the ORM store and deactivate the sequence choice
  - After picking an item from any selector that selector is active and keyboard up and down arrows will change the selection
  - Clone a **Sequence**
  - Component props pass entire objects as much as possible, reduce number of redux queries for performance
  - consider [Dan Abramov's points about Redux](https://stackoverflow.com/questions/35411423/how-to-dispatch-a-redux-action-with-a-timeout/35415559#35415559)
  - consider [redux-saga](https://github.com/redux-saga/redux-saga) -- don't add for now
  - Create a new **Sequence**
  - Create new track in voice that has no tracks
  - Delete a **Sequence**
  - Delete a **Sequence** even if it has children, as long as it has no sequence bindings
  - Delete active **Sequence** with a confirmation
  - Don't reload the entire program after only updating an attribute in ProgramEditor
  - Edit attributes of active **Sequence**
  - Edits properties of **Sequence** (Name, Key, Tempo, Total)
  - Entity child selector (e.g. current Sequence in Program editor, or current Pattern in Voice editor) is a drop-down from a downward chevron only, for maximal efficient use of base screen real estate
  - Fonts and other assets are loaded relative to /mk3/ (currently fonts fail because they are not available at /assets/)
  - Going to route /mk2/program/<id> loads that program from backend API
  - Graceful API error error should prevent user from changing program type of a Rhythm program to any other kind of program, when it has any Tracks and/or Voices.
  - Graceful API error on failed attempt to delete a Pattern
  - Graceful API error when trying to delete Sequence Binding that has Meme
  - Implement [react-redux-toastr](https://www.npmjs.com/package/react-redux-toastr) and errors returned by API during save are displayed in flash messages
  - Implement [react-router](https://www.npmjs.com/package/react-router) and [react-router-dom](https://www.npmjs.com/package/react-router-dom) within new `/mk3` path
  - Implement [redux-api-middleware](https://github.com/agraboso/redux-api-middleware)
  - Implement [redux-orm](https://github.com/redux-orm/redux-orm) and when successful API request is returned update redux-orm store
  - Implement basic [redux](redux.js.org) and do [the tutorial](https://redux.js.org/basics/actions)
  - Macro- and Main-type Programs show a different view of Sequences, including Sequence-Binding CRUD
  - On activating a new Sequence, the pattern choices for all voices are reset to null
  - On creating a new Sequence activate that Sequence
  - Opening the modal to bind a new sequence should focus on the select sequence field
  - React app scaffold
  - React central data store loads from JSON:API
  - redux store slice called `workspace` for persisting the state of the UI workspace
  - Tracks have dotted line background running horizontally through the middle
  - Try building the Stepmatic UI with React
  - use [redux-toolkit](https://redux-toolkit.js.org/)— includes `redux-thunk` and `immer`
  - When UI is loaded and backend is throwing a 502 bad gateway, API should only request once (NOT infinite-loop send requests)

+++ src/main/java/io/xj/core/dao/SegmentDAOImpl.java
@@ -355,8 +355,8 @@ public class SegmentDAOImpl extends DAOImpl<Segment> implements SegmentDAO {
 
     Segment segment = readOne(access, id);
 
-    // Destroy child entities of segment
-    destroyChildEntities(dbProvider.getDSL(), access, id);
+    // Destroy child entities of segment-- but not the messages
+    destroyChildEntities(dbProvider.getDSL(), access, id, false);
 
     update(access, id, segment);
   }
@@ -380,8 +380,8 @@ public class SegmentDAOImpl extends DAOImpl<Segment> implements SegmentDAO {
         waveformKey);
     }
 
-    // Destroy child entities of segment
-    destroyChildEntities(db, access, id);
+    // Destroy ALL child entities of segment
+    destroyChildEntities(db, access, id, true);
 
     // Delete Segment
     db.deleteFrom(SEGMENT)
@@ -397,11 +397,12 @@ public class SegmentDAOImpl extends DAOImpl<Segment> implements SegmentDAO {
   /**
    Destroy all child entities of segment
 
-   @param db     DSL content
-   @param access control
-   @param id     segment to destroy child entities of
+   @param db              DSL content
+   @param access          control
+   @param id              segment to destroy child entities of
+   @param destroyMessages true if we also want to include the segment messages (false to preserve the messages)
    */
-  private void destroyChildEntities(DSLContext db, Access access, UUID id) throws CoreException {
+  private void destroyChildEntities(DSLContext db, Access access, UUID id, Boolean destroyMessages) throws CoreException {
     requireTopLevel(access);
 
     db.deleteFrom(SEGMENT_CHOICE_ARRANGEMENT_PICK)
@@ -424,8 +425,9 @@ public class SegmentDAOImpl extends DAOImpl<Segment> implements SegmentDAO {
       .where(SEGMENT_CHORD.SEGMENT_ID.eq(id))
       .execute();
 
-    db.deleteFrom(SEGMENT_MESSAGE)
-      .where(SEGMENT_MESSAGE.SEGMENT_ID.eq(id))
-      .execute();
+    if (destroyMessages)
+      db.deleteFrom(SEGMENT_MESSAGE)
+        .where(SEGMENT_MESSAGE.SEGMENT_ID.eq(id))
+        .execute();
   }
 }
